Add config versioning
