+++ FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Common/Configurations/SupportedConfigVersion.cs
@@ -1,0 +1,18 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+namespace Microsoft.Health.Fhir.Synapse.Common.Configurations
+{
+    /// <summary>
+    /// Supported versions for configurations.
+    /// </summary>
+    public enum SupportedConfigVersion
+    {
+        /// <summary>
+        /// First version.
+        /// </summary>
+        V1 = 1,
+    }
+}
