+++ FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Tool/appsettings.json
@@ -1,4 +1,5 @@
 {
+  "configVersion": 1,
   "fhirServer": {
     "serverUrl": "",
     "version": "R4",
@@ -22,7 +23,7 @@
   "filter": {
     "enableExternalFilter": false,
     "filterImageReference": "",
-    "filterConfigurationFileName" : "",
+    "filterConfigurationFileName": "",
     "filterScope": "System",
     "groupId": "",
     "requiredTypes": "",
@@ -30,12 +31,12 @@
   },
   "arrow": {
     "readOptions": {
-        "useThreads": true,
-        "blockSize": 30485760
+      "useThreads": true,
+      "blockSize": 30485760
     },
     "writeOptions": {
-        "writeBatchSize": 100,
-        "compression": "SNAPPY"
+      "writeBatchSize": 100,
+      "compression": "SNAPPY"
     }
   },
   "schema": {
